sample_mlb/requirements.txt
streamlit==1.22.0
altair==4.2.0
pandas==1.5.1
numpy==1.23.4
pybase64==1.2.3
matplotlib==3.6.2
seaborn==0.12.1
beautifulsoup4==4.11.1
pillow==9.3.0
requests==2.28.1
lxml==4.9.1
DateTime==5.1